refactor : moved the player statuses to a new component
